Move `rray_multiply_add()` to its own file
