+++ man/rray_multiply_add.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/math-basic.R
+% Please edit documentation in R/multiply-add.R
 \name{rray_multiply_add}
 \alias{rray_multiply_add}
 \title{Fused multiply-add}
@@ -27,7 +27,3 @@ z <- matrix(3L, nrow = 5, ncol = 2)
 x * y + z
 
 }
-\seealso{
-Other math functions: \code{\link{rray_maximum}}
-}
-\concept{math functions}
